refactor(pipes): give builtin prompt migrations their own module
The first pass put prompt copy in two places. A 10-line `concat!` of the
latency preamble lived in mod.rs, the same words lived in the asset, and a
test was supposed to notice if they drifted. A drift test is a worse
answer than a design where drift cannot happen, and it added 135 lines to
a file that was already 11.6k lines long.

Separation, by layer:

  assets/pipes/meeting-summary/pipe.md   prompt wording (the only copy)
  pipes/builtin_migrations.rs            which installs are broken, and why
  pipes/mod.rs                           install/scheduling, calls the above

Replacement text is now sliced out of the shipped asset at call time via
`section_between`, so the migration quotes the prompt instead of copying
it. Verified: rewording only the asset changes what the migration injects,
with no Rust edit. If an anchor ever moves, the slice returns None and the
swap drops out rather than injecting stale wording.

Each repair is a `FragmentSwap { why, old, new }` instead of an anonymous
tuple in a growing match arm. These entries can never be deleted — an
install that has been offline for months still needs them — so the reason
has to live next to the code, not in a commit message. `why` is not
decoration: it is logged when a swap fires, so a support log now says
which defect a given install was carrying.

Moved out of mod.rs with the logic: `replace_prompt_body_when_hash_matches`
(only ever used by migration) and all 9 migration tests, plus 2 new ones
covering `section_between` anchor loss and the derived fast path.

No behavior change. mod.rs is 386 lines lighter — smaller than before this
PR started. 11 migration tests pass, fmt clean, clippy adds no new warning
(the 5 reported are pre-existing in paths.rs), coverage maps up to date.
