+++ crates/screenpipe-core/src/pipes/mod.rs
@@ -10,6 +10,7 @@
 //! parses configs, runs the scheduler, and delegates execution to an
 //! [`AgentExecutor`].
 
+mod builtin_migrations;
 pub mod connection_triggers;
 pub mod connections;
 pub mod favorites;
@@ -23,6 +24,7 @@ use crate::agents::{
     pi::{pi_event_protocol_error, pi_package_enabled, PiExecutor},
     AgentExecutor, ExecutionHandle, SharedPid, STOP_REQUESTED_PID,
 };
+use crate::pipes::builtin_migrations::migrate_builtin_pipe_text;
 use crate::pipes::connections::parse_mcp_connection_id;
 use crate::pipes::mcp_access::McpSessionAccessRegistry;
 use anyhow::{anyhow, Result};
@@ -6611,152 +6613,6 @@ pub fn parse_frontmatter(content: &str) -> Result<(PipeConfig, String)> {
     Ok((config, body))
 }
 
-/// Latency fix for already-installed `meeting-summary` copies.
-///
-/// The shipped prompt told the agent to "read the screenpipe skill first" and
-/// never named the endpoint response shapes, so a real run spent ~8 of its 18
-/// model turns reading skill files and probing JSON shapes (three consecutive
-/// turns just guessing whether `/connections` was an array, an object, or
-/// `.data`). Replacing that one line with the verified shapes plus a tool-call
-/// budget removes the discovery loop without touching the user's own edits.
-const MEETING_SUMMARY_FAST_PATH: &str = concat!(
-    "the user is staring at a spinner until you print step 3, so latency is part of the job. every tool call is a round trip: batch them, and never spend a turn discovering something this prompt already tells you. do not read any skill file — the endpoints and response shapes below are complete and verified. budget: reach step 3 in **6 tool calls or fewer** on a normal meeting.\n",
-    "\n",
-    "these are the exact response shapes. do not probe for them:\n",
-    "\n",
-    "- `GET /meetings/<id>` → a bare object: `{\"id\", \"title\", \"note\", \"meeting_start\", \"meeting_end\", \"meeting_app\", \"attendees\"}`\n",
-    "- `GET /search?...` → `{\"data\": [{\"type\": \"Audio\"|\"OCR\", \"content\": {…}}], \"pagination\": {…}}`\n",
-    "  - audio `content`: `transcription`, `speaker`, `timestamp` (`text` duplicates `transcription`)\n",
-    "  - ocr `content`: `text`, `frame_id`, `app_name`, `window_name`, `timestamp`\n",
-    "- `GET /speakers/unnamed?limit=20&offset=0` → a bare **array** of `{\"id\", \"name\", …}`. `offset` is required; omitting it is a 400.\n",
-    "- `GET /connections` → `{\"data\": [{\"id\", \"name\", \"connected\", \"description\", …}]}` — filter on `connected == true`",
-);
-
-/// Surgically repair known-broken fragments in an already-installed builtin
-/// `pipe.md` without clobbering the user's other edits. `install_builtin_pipes`
-/// only writes when the file is absent, so a stale local copy never picks up a
-/// bundled fix on its own. Rather than overwrite the whole file (which would
-/// throw away user customization), we replace just the specific broken fragment.
-///
-/// Returns the rewritten content only when a migration actually applied, so the
-/// caller can skip the disk write otherwise. Idempotent: running it on
-/// already-fixed content is a no-op.
-fn migrate_builtin_pipe_text(name: &str, original: &str) -> Option<String> {
-    if name == "automate-my-work" {
-        let replacement = BUNDLED_BUILTIN_PIPES
-            .iter()
-            .find_map(|(builtin_name, content)| (*builtin_name == name).then_some(*content))?;
-        return replace_prompt_body_when_hash_matches(
-            original,
-            replacement,
-            AUTOMATE_MY_WORK_LEGACY_PROMPT_HASHES,
-        );
-    }
-
-    // (old, new) fragment swaps per builtin pipe.
-    let replacements: &[(&str, &str)] = match name {
-        // the meeting-summary pipe shipped instructions to PATCH
-        // /meetings/:id, but the server only registers PUT (see
-        // screenpipe-engine server.rs) — so every save 404'd. fix already
-        // installed local copies. PR #4247.
-        "meeting-summary" => &[
-            (
-                "-X PATCH \"http://localhost:3030/meetings/",
-                "-X PUT \"http://localhost:3030/meetings/",
-            ),
-            // the pipe picked "the most recent meeting", which is the wrong one
-            // whenever two meetings end close together. the scheduler now names
-            // the meeting in .trigger-context.json — point installed copies at
-            // it, since install_builtin_pipes never overwrites them. #5481.
-            (
-                "step 1 — find the meeting that just ended:",
-                concat!(
-                    "step 1 — find the meeting that just ended. when the scheduler woke you for an event ",
-                    "it wrote `./.trigger-context.json` in this pipe's folder; read it first and use the ",
-                    "meeting id it names:\n",
-                    "\n",
-                    "  cat ./.trigger-context.json   # {\"event\": \"meeting_ended\", \"key\": \"<MEETING_ID>\", ...}\n",
-                    "\n",
-                    "  curl -s -H \"Authorization: Bearer $SCREENPIPE_LOCAL_API_KEY\" \\\n",
-                    "    \"http://localhost:3030/meetings/<MEETING_ID>\"\n",
-                    "\n",
-                    "only if that file is missing (a manual run) fall back to the most recent row:",
-                ),
-            ),
-            (
-                "the most recent row is the one that just ended. capture its",
-                "either way, capture the meeting's",
-            ),
-            // A maintainer-only synchronization note was accidentally shipped
-            // inside the runtime prompt. Agents interpreted it as work to do
-            // and recursively searched the user's home or ~/.screenpipe for
-            // app source, pegging a CPU core after meetings. Replace it in
-            // already-installed copies with an explicit bounded-work rule.
-            (
-                "keep the wording of this prompt in sync with `buildMeetingSummarizeInstructions` in `apps/screenpipe-app-tauri/lib/utils/meeting-context.ts` (used by the in-app \"summarize with AI\" button) — the two surfaces should produce the same behavior.",
-                "the instructions below are complete. screenpipe API search is required: use the meeting id and exact meeting time window with the named local HTTP endpoints below. do not inspect app source or recursively search the filesystem; never run recursive `find` or `grep` over the user's home or `~/.screenpipe`.",
-            ),
-            // Clarify the first bounded-work migration: the pipe must search
-            // Screenpipe through the meeting-scoped API. Only recursive
-            // filesystem/source discovery is prohibited.
-            (
-                "the instructions below are complete. do not inspect app source or search outside this pipe folder. never run recursive `find` or `grep` over the user's home or `~/.screenpipe`; use only the named local files and bounded HTTP endpoints below.",
-                "the instructions below are complete. screenpipe API search is required: use the meeting id and exact meeting time window with the named local HTTP endpoints below. do not inspect app source or recursively search the filesystem; never run recursive `find` or `grep` over the user's home or `~/.screenpipe`.",
-            ),
-            // `/speakers/unnamed` rejects a request without `offset` (400
-            // "missing field `offset`"), so the shipped call failed on every
-            // run and cost a recovery round trip. Pin the working query.
-            (
-                "\"http://localhost:3030/speakers/unnamed?limit=20\"",
-                "\"http://localhost:3030/speakers/unnamed?limit=20&offset=0\"",
-            ),
-            // Latency: the pipe blocks a visible spinner, but shipped
-            // instructions made the agent read two skill files and discover
-            // endpoint shapes it was never told, costing ~8 of 18 turns on a
-            // 7-minute meeting. Tell it the shapes and give it a budget.
-            (
-                "read the screenpipe skill first so you know the meetings + search endpoints.",
-                MEETING_SUMMARY_FAST_PATH,
-            ),
-        ],
-        _ => return None,
-    };
-
-    let mut updated = original.to_string();
-    for (old, new) in replacements {
-        updated = updated.replace(old, new);
-    }
-
-    (updated != original).then_some(updated)
-}
-
-/// Replace only the instruction body of a known built-in prompt version.
-/// Frontmatter and the self-improving memory section stay untouched. Any user
-/// edit inside the instruction body changes the hash and opts out of migration.
-fn replace_prompt_body_when_hash_matches(
-    original: &str,
-    replacement: &str,
-    legacy_hashes: &[&str],
-) -> Option<String> {
-    let original_prompt_start = original.find("<role>")?;
-    let replacement_prompt_start = replacement.find("<role>")?;
-    let original_prompt = &original[original_prompt_start..];
-    let original_hash = simple_hash(original_prompt);
-    if !legacy_hashes.contains(&original_hash.as_str()) {
-        return None;
-    }
-
-    let prefix = original[..original_prompt_start].replace(
-        "description: \"Find genuinely new, low-risk automations tailored to your workflow\"",
-        "description: \"Find one repeated workflow and propose a testable automation\"",
-    );
-    Some(format!(
-        "{}{}",
-        prefix,
-        &replacement[replacement_prompt_start..]
-    ))
-}
-
 /// Atomic file write: write to a temp file in the same directory, then rename.
 /// On Unix, rename is atomic. On Windows, this avoids the partial-write window
 /// where a concurrent reader (e.g. the scheduler) sees a truncated file.
@@ -8767,246 +8623,6 @@ mod tests {
         assert!(logs_dir
             .join("archive/2026-01/20260101_000000.json")
             .exists());
-    }
-
-    #[test]
-    fn migrate_builtin_pipe_fixes_stale_patch_verb() {
-        // a stale local meeting-summary copy with the old PATCH verb gets
-        // surgically rewritten to PUT, leaving surrounding text untouched.
-        let stale = "do stuff\n  curl -s -X PATCH \"http://localhost:3030/meetings/<MEETING_ID>\" \\\n    -d '{}'\nmore stuff";
-        let fixed = migrate_builtin_pipe_text("meeting-summary", stale)
-            .expect("stale PATCH content should migrate");
-        assert!(fixed.contains("-X PUT \"http://localhost:3030/meetings/"));
-        assert!(!fixed.contains("-X PATCH"));
-        assert!(fixed.starts_with("do stuff"));
-        assert!(fixed.ends_with("more stuff"));
-
-        // idempotent: already-PUT content is a no-op (no rewrite, no churn).
-        assert!(migrate_builtin_pipe_text("meeting-summary", &fixed).is_none());
-
-        // other builtins and unrelated content are left alone.
-        assert!(migrate_builtin_pipe_text("day-recap", stale).is_none());
-        assert!(migrate_builtin_pipe_text("meeting-summary", "no api calls here").is_none());
-    }
-
-    /// #5481: installed copies still say "pick the most recent meeting", which is
-    /// wrong when two meetings end together. They must be pointed at the trigger
-    /// file, since install_builtin_pipes never overwrites an existing pipe.md.
-    #[test]
-    fn migrate_builtin_pipe_points_meeting_summary_at_trigger_context() {
-        let stale = concat!(
-            "read the screenpipe skill first.\n",
-            "\n",
-            "step 1 — find the meeting that just ended:\n",
-            "\n",
-            "  curl -s \"http://localhost:3030/meetings?limit=1\"\n",
-            "\n",
-            "the most recent row is the one that just ended. capture its `id` and `note`.\n",
-            "\n",
-            "step 2 — summarize it.",
-        );
-
-        let fixed = migrate_builtin_pipe_text("meeting-summary", stale)
-            .expect("stale most-recent-meeting content should migrate");
-        assert!(fixed.contains(".trigger-context.json"));
-        assert!(fixed.contains("/meetings/<MEETING_ID>"));
-        assert!(!fixed.contains("the most recent row is the one that just ended"));
-
-        // the manual-run fallback and the surrounding steps survive.
-        assert!(fixed.contains("http://localhost:3030/meetings?limit=1"));
-        assert!(fixed.starts_with("read the screenpipe skill first."));
-        assert!(fixed.ends_with("step 2 — summarize it."));
-
-        // idempotent: running it again is a no-op.
-        assert!(migrate_builtin_pipe_text("meeting-summary", &fixed).is_none());
-    }
-
-    #[test]
-    fn migrate_builtin_pipe_removes_source_search_instruction() {
-        let stale = concat!(
-            "a meeting just ended.\n\n",
-            "keep the wording of this prompt in sync with `buildMeetingSummarizeInstructions` in ",
-            "`apps/screenpipe-app-tauri/lib/utils/meeting-context.ts` (used by the in-app ",
-            "\"summarize with AI\" button) — the two surfaces should produce the same behavior.\n\n",
-            "read the screenpipe skill first.\n",
-        );
-
-        let fixed = migrate_builtin_pipe_text("meeting-summary", stale)
-            .expect("source-search instruction should migrate");
-        assert!(!fixed.contains("buildMeetingSummarizeInstructions"));
-        assert!(fixed.contains("screenpipe API search is required"));
-        assert!(fixed.contains("never run recursive `find` or `grep`"));
-        assert!(fixed.ends_with("read the screenpipe skill first.\n"));
-        assert!(migrate_builtin_pipe_text("meeting-summary", &fixed).is_none());
-    }
-
-    /// The shipped prompt must already be in its migrated form, or every fresh
-    /// install would be rewritten on the next startup.
-    #[test]
-    fn bundled_meeting_summary_needs_no_migration() {
-        let bundled = BUNDLED_BUILTIN_PIPES
-            .iter()
-            .find_map(|(name, content)| (*name == "meeting-summary").then_some(*content))
-            .expect("meeting-summary is bundled");
-        assert!(migrate_builtin_pipe_text("meeting-summary", bundled).is_none());
-        let (config, body) = parse_frontmatter(bundled).expect("bundled prompt should parse");
-        assert_eq!(config.timeout, Some(600));
-        assert!(!body.contains("buildMeetingSummarizeInstructions"));
-        assert!(body.contains("screenpipe API search is required"));
-        assert!(body.contains("never run recursive `find` or `grep`"));
-    }
-
-    /// `/speakers/unnamed` 400s without `offset`, so the shipped query failed on
-    /// every run and the agent burned a turn recovering from it.
-    #[test]
-    fn migrate_builtin_pipe_fixes_unnamed_speakers_offset() {
-        let stale = concat!(
-            "step 2d — name the speakers.\n",
-            "  curl -s -H \"Authorization: Bearer $SCREENPIPE_LOCAL_API_KEY\" \\\n",
-            "    \"http://localhost:3030/speakers/unnamed?limit=20\"\n",
-            "only rename when the evidence is unambiguous.\n",
-        );
-
-        let fixed = migrate_builtin_pipe_text("meeting-summary", stale)
-            .expect("a query missing `offset` should migrate");
-        assert!(fixed.contains("/speakers/unnamed?limit=20&offset=0"));
-        assert!(!fixed.contains("/speakers/unnamed?limit=20\""));
-        // the user's surrounding prose is untouched.
-        assert!(fixed.ends_with("only rename when the evidence is unambiguous.\n"));
-        assert!(migrate_builtin_pipe_text("meeting-summary", &fixed).is_none());
-    }
-
-    /// Latency: the shipped prompt sent the agent to read skill files and left it
-    /// to probe endpoint shapes, which cost roughly eight of eighteen turns on a
-    /// real 7-minute meeting while the user watched a spinner.
-    #[test]
-    fn migrate_builtin_pipe_gives_meeting_summary_a_tool_call_budget() {
-        let stale = concat!(
-            "a meeting just ended.\n",
-            "\n",
-            "read the screenpipe skill first so you know the meetings + search endpoints.\n",
-            "\n",
-            "step 1 — find the meeting that just ended.\n",
-            "my own note: always cc me.\n",
-        );
-
-        let fixed = migrate_builtin_pipe_text("meeting-summary", stale)
-            .expect("skill-read instruction should migrate");
-
-        // the discovery loop is gone.
-        assert!(!fixed.contains("read the screenpipe skill first"));
-        assert!(fixed.contains("do not read any skill file"));
-        // and it is replaced by a budget plus the shapes it used to guess at.
-        assert!(fixed.contains("6 tool calls or fewer"));
-        assert!(fixed.contains("these are the exact response shapes"));
-        assert!(fixed.contains("`offset` is required"));
-        assert!(fixed.contains("filter on `connected == true`"));
-
-        // the user's own edits and the surrounding steps survive.
-        assert!(fixed.starts_with("a meeting just ended.\n"));
-        assert!(fixed.ends_with("my own note: always cc me.\n"));
-        assert!(fixed.contains("step 1 — find the meeting that just ended."));
-
-        assert!(migrate_builtin_pipe_text("meeting-summary", &fixed).is_none());
-    }
-
-    /// The migration text and the shipped prompt must stay one source of truth.
-    /// If they drift, installed copies get advice the bundled prompt contradicts.
-    #[test]
-    fn meeting_summary_fast_path_matches_bundled_prompt() {
-        let bundled = BUNDLED_BUILTIN_PIPES
-            .iter()
-            .find_map(|(name, content)| (*name == "meeting-summary").then_some(*content))
-            .expect("meeting-summary is bundled");
-        assert!(
-            bundled.contains(MEETING_SUMMARY_FAST_PATH),
-            "MEETING_SUMMARY_FAST_PATH drifted from the bundled prompt"
-        );
-    }
-
-    /// The shipped prompt must fetch in parallel rather than one endpoint per
-    /// turn, and must not reintroduce the round trips this pipe just removed.
-    #[test]
-    fn bundled_meeting_summary_batches_its_fetches() {
-        let bundled = BUNDLED_BUILTIN_PIPES
-            .iter()
-            .find_map(|(name, content)| (*name == "meeting-summary").then_some(*content))
-            .expect("meeting-summary is bundled");
-        let (_, body) = parse_frontmatter(bundled).expect("bundled prompt should parse");
-
-        // one batched fetch, backgrounded and joined.
-        assert!(body.contains("pull everything the summary needs in ONE command"));
-        assert!(body.contains("-o /tmp/audio.json &"));
-        assert!(body.contains("-o /tmp/ocr.json &"));
-        assert!(body.contains("\n  wait\n"));
-
-        // bounded up front, so there is no "too big, fetch again" round trip.
-        assert!(body.contains("do not fetch unbounded and then re-fetch smaller"));
-
-        // the endpoints reused from that batch are not re-fetched later.
-        assert!(body.contains("already fetched to /tmp/spk.json in step 1"));
-        assert!(body.contains("/tmp/conn.json"));
-
-        // and the discovery loop stays gone.
-        assert!(!body.contains("read the screenpipe skill first"));
-        assert!(body.contains("6 tool calls or fewer"));
-
-        // the streaming contract the UI depends on is preserved.
-        assert!(body.contains("## Summary"));
-    }
-
-    #[test]
-    fn migrate_builtin_pipe_replaces_only_a_known_prompt_body() {
-        let stale = concat!(
-            "---\nschedule: manual\n",
-            "description: \"Find genuinely new, low-risk automations tailored to your workflow\"\n",
-            "---\n\n# memory\n- user lesson\n\n",
-            "<role>\nlegacy automation instructions\n</role>\n",
-        );
-        let replacement = concat!(
-            "---\nschedule: manual\n---\n\n",
-            "<role>\nnew evidence-first instructions\n</role>\n",
-        );
-        let prompt_start = stale.find("<role>").unwrap();
-        let legacy_hash = simple_hash(&stale[prompt_start..]);
-
-        let fixed =
-            replace_prompt_body_when_hash_matches(stale, replacement, &[legacy_hash.as_str()])
-                .expect("known legacy prompt should migrate");
-
-        assert!(fixed.contains("# memory\n- user lesson"));
-        assert!(fixed.contains(
-            "description: \"Find one repeated workflow and propose a testable automation\""
-        ));
-        assert!(fixed.contains("new evidence-first instructions"));
-        assert!(!fixed.contains("legacy automation instructions"));
-        assert!(replace_prompt_body_when_hash_matches(
-            &fixed,
-            replacement,
-            &[legacy_hash.as_str()],
-        )
-        .is_none());
-
-        let customized = stale.replace(
-            "legacy automation instructions",
-            "legacy automation instructions with my customization",
-        );
-        assert!(replace_prompt_body_when_hash_matches(
-            &customized,
-            replacement,
-            &[legacy_hash.as_str()],
-        )
-        .is_none());
-    }
-
-    #[test]
-    fn current_automate_my_work_builtin_does_not_migrate_again() {
-        let current = BUNDLED_BUILTIN_PIPES
-            .iter()
-            .find_map(|(name, content)| (*name == "automate-my-work").then_some(*content))
-            .unwrap();
-
-        assert!(migrate_builtin_pipe_text("automate-my-work", current).is_none());
     }
 
     #[test]
